feat: an approach to the reader separation
